+++ ProdManager/helpers/date.py
@@ -2,6 +2,10 @@ from datetime import datetime
 
 from flask import g
 
+DATE_FORMAT = '%Y-%m-%d'
+DATETIME_FORMAT = '%Y-%m-%dT%H:%M'
+READABLE_DATETIME_FORMAT = '%d/%m/%Y %H:%M'
+
 def current_date(rounded=True):
   date = datetime.now()
 
@@ -12,6 +16,6 @@ def current_date(rounded=True):
 
 def beautifull_date(date):
   if g.api:
-    return date.strftime('%Y-%m-%dT%H:%M')
+    return date.strftime(DATETIME_FORMAT)
 
-  return date.strftime('%d/%m/%Y %H:%M')
+  return date.strftime(READABLE_DATETIME_FORMAT)